Committing work  to date

+++ src/components/table/MyStudentTable.js
@@ -1,7 +1,13 @@
+/* eslint-disable no-alert */
+/* eslint-disable react-hooks/exhaustive-deps */
 import React, { useEffect, useState } from 'react'
 import './table.scss'
+import './buttons.scss'
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Link } from 'react-router-dom'
+import {
+  Trash2, Edit, PlusSquare, PlusCircle, Maximize2, MinusCircle, X
+} from 'react-feather'
 
 function MyTable() {
   const headerCols = [
@@ -9,77 +15,118 @@ function MyTable() {
     'Forename',
     'Surname',
     'Enrolment Year',
-    'Graduation Year'
+    'Graduation Year',
+    'Actions'
   ]
-  //  const mainData = [...]
   const [mainData, setMainData] = useState([])
   const baseUrl = 'http://localhost:8080/api/students'
   const [myUrl, setMyUrl] = useState(baseUrl)
   const [selectFilter, setSelectFilter] = useState('studentId')
   const [filterValue, setFilterValue] = useState('')
   const [isFilterAdded, setIsFilterAdded] = useState(false)
-  const [errorMessage, setErrorMessage] = useState('')
-  const [infoMessage, setInfoMessage] = useState('')
-  const [isMessageVisible, setIsMessageVisible] = React.useState(false)
   const [newForename, setNewForename] = useState('')
   const [newSurname, setNewSurname] = useState('')
   const [newEnrolmentYear, setNewEnrolmentYear] = useState('')
   const [newGraduationYear, setNewGraduationYear] = useState('')
-  const [isNewStudentAdded, setIsNewStudentAdded] = useState(false)
   const [editingRow, setEditingRow] = useState()
   const [isDataRefreshNeeded, setIsDataRefreshNeeded] = useState(false)
-  // const [isStudentDeleted, setIsStudentDeleted] = useState(false)
-  setTimeout(() => {
-    setIsMessageVisible(false)
-    console.log('disappear')
-  }, 10000)
+  const [isAddStudentFormVisible, setIsAddStudentFormVisible] = useState(false)
+  const [postFailed, setPostFailed] = useState(false)
+
+  const [filterMessage, setFilterMessage] = useState('')
+  const [actionMessage, setActionMessage] = useState('')
+  const [isFilterMessageVisible, setIsFilterMessageVisible] = React.useState(false)
+  const [isActionMessageVisible, setIsActionMessageVisible] = React.useState(false)
+  const [filterMessageTimeoutHandle, setFilterMessageTimeoutHandle] = useState(0)
+  const [actionMessageTimeoutHandle, setActionMessageTimeoutHandle] = useState(0)
+
+  const [isError, setIsError] = useState(false)
+  function resetInputRowValues() {
+    setNewForename('')
+    setNewSurname('')
+    setNewGraduationYear('')
+    setNewEnrolmentYear('')
+    setPostFailed(false)
+  }
+  function toggleAddStudentForm() {
+    setIsAddStudentFormVisible(!isAddStudentFormVisible)
+    resetInputRowValues()
+  }
+  const filterMessageTimeout = () => {
+    setFilterMessageTimeoutHandle(setTimeout(() => {
+      setIsFilterMessageVisible(false)
+      setFilterMessage('')
+    }, 5000))
+  }
+  const actionMessageTimeout = () => {
+    setActionMessageTimeoutHandle(setTimeout(() => {
+      setIsActionMessageVisible(false)
+      setActionMessage('')
+    }, 5000))
+  }
+  const showMessage = (message, type) => {
+    if (type === 'filtermessage') {
+      clearTimeout(filterMessageTimeoutHandle)
+      setFilterMessage(message)
+      setIsFilterMessageVisible(true)
+      filterMessageTimeout()
+    } else if (type === 'actionerror') {
+      setIsError(true)
+      clearTimeout(actionMessageTimeoutHandle)
+      setActionMessage(message)
+      setIsActionMessageVisible(true)
+      actionMessageTimeout()
+    } else if (type === 'actionsuccess') {
+      setIsError(false)
+      clearTimeout(actionMessageTimeoutHandle)
+      setActionMessage(message)
+      setIsActionMessageVisible(true)
+      actionMessageTimeout()
+    }
+  }
+  const displayActionMessage = (
+    <text className="actionmessage" style={{ fontWeight: 'bold', color: isError ? 'red' : 'royalblue' }}>
+      {actionMessage}
+    </text>
+  )
+  const displayFilterMessage = (
+    <div className="filtermessage">
+      <text>
+        {filterMessage}
+      </text>
+    </div>
+  )
   const resetTableData = () => {
     setMyUrl(baseUrl)
     setIsDataRefreshNeeded(true)
+    setIsAddStudentFormVisible(false)
   }
   useEffect(() => {
-    console.log('I am using effect hook')
-
     fetch(myUrl)
       .then((response) => response.json())
       .then((data) => {
-        console.log('datareceived:', data)
-
         if (typeof (data.status) === 'undefined') {
           setMainData(data)
-          setErrorMessage('')
           if (data.length === 0) {
-            setInfoMessage('No data found')
+            showMessage('No data found', 'filtermessage')
           }
         } else {
-          console.log('data.status', data.status)
-
-          setMainData('error')
-          setErrorMessage(data.message)
+          showMessage(data.message, 'filtermessage')
+          setMyUrl(baseUrl)
         }
         setIsFilterAdded(false)
-        setIsNewStudentAdded(false)
         setFilterValue('')
         setIsDataRefreshNeeded(false)
       })
-  }, [myUrl, isFilterAdded, isNewStudentAdded, isDataRefreshNeeded])
+  }, [myUrl, isFilterAdded, isDataRefreshNeeded])
 
   const updateRow = (value, rowData, field) => {
     const rowToUpdate = mainData.filter((row) => (row.studentId === rowData.studentId))[0]
-    console.log('value', value)
-    console.log('field', field)
-    console.log('rowToUpdate', rowToUpdate)
-
     const prevValue = rowData[field]
     rowToUpdate[field] = value
     const student = rowData
     student[field] = value
-    console.log('student', student)
-
-    console.log('prevVlaue', prevValue)
-    console.log('value', rowToUpdate[field])
     if (rowData[field] !== prevValue) {
-      console.log('rowToUpdate', rowToUpdate)
       fetch(`${baseUrl}/${rowToUpdate.studentId}`, {
         method: 'PUT',
         headers: {
@@ -90,14 +137,11 @@ function MyTable() {
         .then((response) => {
           if (response.ok) {
             rowToUpdate[field] = value
-            setInfoMessage('Student updated')
-            setIsMessageVisible(true)
+            showMessage(`Student with ID: ${rowData.studentId} updated successfully`, 'actionsuccess')
           } else {
-            console.log('response', response)
             rowToUpdate[field] = prevValue
             setEditingRow([])
-            setInfoMessage('Error updating student')
-            setIsMessageVisible(true)
+            showMessage('Error updating student, please try again', 'actionerror')
           }
         })
     }
@@ -105,54 +149,44 @@ function MyTable() {
   }
 
   const removeRow = (rowData) => {
-    console.log('removeRow', rowData)
-    console.log('filter: ', mainData.filter((row) => (row.studentId !== rowData)))
-    setMainData(mainData.filter((row) => (row.studentId !== rowData)))
+    if (window.confirm('Are you sure you want to delete this student?')) {
+      setMainData(mainData.filter((row) => (row.studentId !== rowData)))
 
-    fetch(`${baseUrl}/${rowData}`, { method: 'DELETE' })
-      .then((response) => {
-        console.log('delete data received: ', response)
-        if (response.status === 200) {
-          setInfoMessage(`Studentwith ID: ${rowData} deleted`)
-          setIsMessageVisible(true)
-        } else {
-          setInfoMessage('Error deleting student')
-          setIsMessageVisible(true)
-        }
-      })
+      fetch(`${baseUrl}/${rowData}`, { method: 'DELETE' })
+        .then((response) => {
+          if (response.status === 200) {
+            showMessage(`Student with ID: ${rowData} deleted`, 'actionsuccess')
+          } else {
+            showMessage('Error deleting student', 'actionerror')
+          }
+        })
+    }
   }
+
   const handlePostStudent = (event) => {
     event.preventDefault()
-    const newStudent = {
-      forename: `${newForename}`,
-      surname: `${newSurname}`,
-      enrolmentYear: `${newEnrolmentYear}`,
-      graduationYear: `${newGraduationYear}`
+    if (window.confirm('Are you sure you want to add this student?')) {
+      const newStudent = {
+        forename: `${newForename}`,
+        surname: `${newSurname}`,
+        enrolmentYear: `${newEnrolmentYear}`,
+        graduationYear: `${newGraduationYear}`
+      }
+      fetch(baseUrl, { method: 'POST', body: JSON.stringify(newStudent), headers: { 'Content-Type': 'application/json' } })
+        .then((response) => {
+          if (response.status === 200) {
+            showMessage(`Student: ${newForename} ${newSurname} added`, 'actionsuccess')
+            resetInputRowValues()
+            setIsDataRefreshNeeded(true)
+          } else {
+            showMessage('Error adding student: Please check the data', 'actionerror')
+            setPostFailed(true)
+          }
+        })
     }
-    console.log('newStudent', newStudent)
-    console.log('newStudent', JSON.stringify(newStudent))
-    fetch(baseUrl, { method: 'POST', body: JSON.stringify(newStudent), headers: { 'Content-Type': 'application/json' } })
-      .then((response) => {
-        console.log('data received: ', response)
-        if (response.status === 200) {
-          setInfoMessage(`Student: ${newForename} ${newSurname} added`)
-          setIsMessageVisible(true)
-          setIsNewStudentAdded(true)
-        } else {
-          setInfoMessage('Error adding student: Please check the data')
-          setIsMessageVisible(true)
-        }
-        setNewForename('')
-        setNewSurname('')
-        setNewGraduationYear('')
-        setNewEnrolmentYear('')
-      })
-    // .then(() => setMainData(() => [...mainData, newStudent]))
-    console.log('list', mainData)
   }
 
   const applySearchFilter = () => {
-    console.log('handlingfilter')
     if (filterValue !== '') {
       if (selectFilter === 'studentId') {
         setMyUrl(`${baseUrl}/${filterValue}`)
@@ -167,24 +201,49 @@ function MyTable() {
       }
     } else {
       resetTableData()
+      showMessage('Please enter a value to filter', 'filtermessage')
+    }
+  }
+  const handleEditRow = (dataId) => {
+    if (dataId === editingRow) {
+      setEditingRow()
+    } else {
+      setEditingRow(dataId)
     }
   }
   const filter = (
-    <div>
-      <select value={selectFilter} onChange={(e) => setSelectFilter(e.currentTarget.value)}>
+    <div className="filter">
+      <select className="dropdown" value={selectFilter} onChange={(e) => setSelectFilter(e.currentTarget.value)}>
         <option value="studentId">Student ID</option>
         <option value="name">Student Name</option>
         <option value="semester">Semester</option>
       </select>
-      <label>
-        Search:
-        <input type="text" name="filterValue" value={filterValue} onChange={(e) => setFilterValue(e.currentTarget.value)} />
-        <button type="button" onClick={applySearchFilter}>
+
+      <label className="search">
+        Filter Value:
+        <input className="searchbar" type="text" name="filterValue" value={filterValue} onChange={(e) => setFilterValue(e.currentTarget.value)} />
+        <button className="filterbutton" type="button" onClick={applySearchFilter}>
           Filter
         </button>
+
       </label>
+      {isFilterMessageVisible
+  && displayFilterMessage}
+
+      {!isAddStudentFormVisible ? (
+        <button className="buttonaddnew" style={{ marginLeft: 'auto' }} type="button" onClick={() => { toggleAddStudentForm() }}>
+          New Student&nbsp;
+          <PlusCircle size="1.7vw" />
+        </button>
+      ) : (
+        <button className="buttonaddnew" style={{ marginLeft: 'auto' }} type="button" onClick={() => { toggleAddStudentForm() }}>
+          New Student&nbsp;
+          <MinusCircle size="1.7vw" />
+        </button>
+      )}
     </div>
   )
+
   return (
     <div>
       {filter}
@@ -193,19 +252,21 @@ function MyTable() {
           <tr>
             {' '}
             {headerCols.map((col) => (
-              <td>
+              <th>
                 {col}
-              </td>
+              </th>
             ))}
           </tr>
-
         </thead>
         <tbody>
-          {errorMessage === '' ? mainData.map((data) => (
+          {mainData.map((data) => (
             <tr key={data.studentId}>
               {Object.entries(data).map(([prop, value]) => (
                 <td
-                  contentEditable={data.studentId === editingRow}
+                  style={{
+                    border: (data.studentId === editingRow) && (prop !== 'studentId') ? '2px solid blue' : '1px solid black'
+                  }}
+                  contentEditable={(data.studentId === editingRow) && (prop !== 'studentId')}
                   // eslint-disable-next-line react/no-unknown-property
                   field={prop}
                   onBlur={(event) => {
@@ -216,57 +277,89 @@ function MyTable() {
                 </td>
               ))}
               <td className="tdinvisible">
-                <button className="buttonedit" type="button" onClick={() => { setEditingRow(data.studentId) }}>
-                  Edit Row
+                <button className="buttonedit-students" type="button" onClick={() => { handleEditRow(data.studentId) }}>
+                  <Edit className="icon-sizing" />
                 </button>
-                <button className="buttondelete" type="button" onClick={() => { removeRow(data.studentId) }}>
-                  Delete Row
+                <button className="buttondelete-students" type="button" onClick={() => { removeRow(data.studentId) }}>
+                  <Trash2 className="icon-sizing" />
                 </button>
-              </td>
-              <td className="tdinvisible">
                 <Link to={`/enrolment/${data.studentId}`}>
-                  <button className="buttonenroll" type="button">
-                    View Enrolled Courses
+                  <button className="buttonviewenroll-students" type="button">
+                    <Maximize2 className="icon-sizing" />
                   </button>
                 </Link>
               </td>
             </tr>
-          )) : <tr><td className="error">{errorMessage}</td></tr>}
-          <tr>
+          ))}
+          {isAddStudentFormVisible && (
+          <tr style={postFailed ? { backgroundColor: 'darkred' } : { backgroundColor: 'steelblue' }}>
             <td>New</td>
             <td>
-              <input type="text" name="forename" onChange={(e) => setNewForename(e.currentTarget.value)} value={newForename} />
+              <div>
+                <input
+                  size="1"
+                  type="text"
+                  name="forename"
+                  onChange={(e) => setNewForename(e.currentTarget.value)}
+                  value={newForename}
+                  onClick={() => setPostFailed(false)}
+                />
+              </div>
             </td>
             <td>
-              <input type="text" name="surname" onChange={(e) => setNewSurname(e.currentTarget.value)} value={newSurname} />
+              <div>
+                <input
+                  size="1"
+                  type="text"
+                  name="surname"
+                  onChange={(e) => setNewSurname(e.currentTarget.value)}
+                  value={newSurname}
+                  onClick={() => setPostFailed(false)}
+                />
+              </div>
             </td>
             <td>
-              <input type="text" name="enrolmentyear" onChange={(e) => setNewEnrolmentYear(e.currentTarget.value)} value={newEnrolmentYear} />
+              <div>
+                <input
+                  size="1"
+                  type="text"
+                  name="enrolmentyear"
+                  onChange={(e) => setNewEnrolmentYear(e.currentTarget.value)}
+                  value={newEnrolmentYear}
+                  onClick={() => setPostFailed(false)}
+                />
+              </div>
             </td>
             <td>
-              <input type="text" name="graduationYear" onChange={(e) => setNewGraduationYear(e.currentTarget.value)} value={newGraduationYear} />
+              <div>
+                <input
+                  size="1"
+                  type="text"
+                  name="graduationYear"
+                  onChange={(e) => setNewGraduationYear(e.currentTarget.value)}
+                  value={newGraduationYear}
+                  onClick={() => setPostFailed(false)}
+                />
+              </div>
             </td>
             <td className="tdinvisible">
-              <button className="buttonadd" type="button" onClick={(e) => { handlePostStudent(e) }}>
-                Add Student
+              <button className="buttonadd-students" type="button" onClick={(e) => { handlePostStudent(e) }}>
+                <PlusSquare className="icon-sizing" />
+              </button>
+              <button className="buttondelete-students" type="button" onClick={() => { resetInputRowValues() }}>
+                <X className="icon-sizing" />
               </button>
             </td>
           </tr>
-
+          )}
         </tbody>
 
       </table>
-      <button type="button" onClick={resetTableData}>
+      <button className="resetbutton" type="button" onClick={resetTableData}>
         Reset Table
       </button>
-      {isMessageVisible
-        && (
-          <h3 className="error">
-            {' '}
-            {infoMessage}
-            {' '}
-          </h3>
-        )}
+      {isActionMessageVisible
+        && displayActionMessage}
     </div>
   )
 }
